backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
asttokens==2.4.1
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.8
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
entrypoints==0.4
executing==2.1.0
fastapi==0.115.5
fasteners==0.14.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
fonttools==4.55.0
fs==2.4.12
fsspec==2024.10.0
future==0.18.2
gast==0.5.2
gensim==4.3.3
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
icoextract==0.1.2
idna==3.10
importlib-metadata==4.6.4
IProgress==0.4
ipykernel==6.29.5
ipython==8.29.0
ipython_genutils==0.2.0
ipywidgets==8.1.5
jedi==0.19.2
jeepney==0.7.1
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==3.2.0
jupyter-console==6.4.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab-pygments==0.1.2
jupyterlab_widgets==3.0.13
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
lxml_html_clean==0.4.1
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
notebook==6.4.8
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.4
pefile==2021.9.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.6
ply==3.11
prometheus-client==0.9.0
prompt_toolkit==3.0.48
protobuf==3.12.4
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.2
pydantic_core==2.27.1
pyelftools==0.27
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt===2.4.0-ubuntu4-zorin1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2024.2
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.5.2
scipy==1.13.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
setuptools==75.5.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.2.3
uvicorn==0.34.0
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zorin-appearance==5.3
zorin-connect==1.2
zorin-exec-guard==1.5
